+++ src/presentation/adapters/input_book_one_create.go
@@ -6,7 +6,7 @@ import (
 )
 
 type BookOneCreateInput struct {
-	GenreID     int64      `json:"genreId"`
+	GenreID     int        `json:"genreId"`
 	Title       string     `json:"title"`
 	Description string     `json:"description"`
 	Price       PriceInput `json:"price"`
@@ -18,7 +18,7 @@ func (r *BookOneCreateInput) ToEntity() *domain.Book {
 		Title:       r.Title,
 		Description: r.Description,
 		Price: domain.Price{
-			Amount:   uint64(r.Price.Amount),
+			Amount:   uint(r.Price.Amount),
 			Currency: r.Price.Currency,
 		},
 		CreatedAt: time.Now(),
