+++ src/presentation/adapters/input_book_one_update.go
@@ -7,7 +7,7 @@ import (
 
 type BookOneUpdateInput struct {
 	ID          int        `json:"id"`
-	GenreID     int64      `json:"genreId"`
+	GenreID     int        `json:"genreId"`
 	Title       string     `json:"title"`
 	Description string     `json:"description"`
 	Price       PriceInput `json:"price"`
@@ -15,12 +15,12 @@ type BookOneUpdateInput struct {
 
 func (r *BookOneUpdateInput) ToEntity() *domain.Book {
 	res := &domain.Book{
-		ID:          int64(r.ID),
+		ID:          r.ID,
 		GenreID:     r.GenreID,
 		Title:       r.Title,
 		Description: r.Description,
 		Price: domain.Price{
-			Amount:   uint64(r.Price.Amount),
+			Amount:   uint(r.Price.Amount),
 			Currency: r.Price.Currency,
 		},
 		UpdatedAt: time.Now(),
